// src/types/user.ts
export interface IUserNew {
  id?: string;
  userName: string;
  userEmail: string;
  userPass: string;
  userPhone: string;
  userGender: string;
  userAdress: string;
  userRole: string;
}

export interface IUser {
  id?: string;
  fullNameUser?: string;
  emailUser?: string;
  passwordUser?: string;
  phoneUser?: string;
  genderUser?: string;
  adressUser?: string;
  typeRole?: string;
}